Fixinf

+++ app/Http/Controllers/PostController.php
@@ -71,7 +71,9 @@ class PostController extends Controller
         // Replace the collection in the paginator instance
         $posts->setCollection($mergedPosts->values());
 
-        $badges = User::with('badges')->get();
+        $badges = User::with(['badges' => function($q) {
+            $q->orderBy('point', 'asc');
+        }])->get();
 
 
         return view('home', [
@@ -481,6 +483,7 @@ class PostController extends Controller
 
         foreach ($post->attachments as $attachment) {
             if ($attachment->namafile === 'blankimage.png') continue;
+            if (!$attachment->imgkit_id) continue; // Skip if imgkit_id is not set
             $hslnya = $imageKit->deleteFile($attachment->imgkit_id);
             if ($hslnya->error) {
                 return response()->json([
